readme tweaks, other tweaks for agent context.
